Flatten Polars
This both brings polars into a single file, and it leverages
coiled.functions a bit more explicitly in tests.  I *think* that this is
more readable, and it also separates restart times from execution times.
