--- tests/benchmarks/tpch/polars_queries/__init__.py
@@ -1,1 +1,0 @@
-from . import q1, q2, q3, q4, q5, q6, q7  # noqa: F401
