--- tests/benchmarks/tpch/polars_queries/q3.py
@@ -1,37 +1,0 @@
-from datetime import datetime
-
-import polars as pl
-
-from .utils import read_data
-
-
-def query():
-    var_1 = var_2 = datetime(1995, 3, 15)
-    var_3 = "BUILDING"
-
-    customer_ds = read_data("customer")
-    line_item_ds = read_data("lineitem")
-    orders_ds = read_data("orders")
-
-    return (
-        customer_ds.filter(pl.col("c_mktsegment") == var_3)
-        .join(orders_ds, left_on="c_custkey", right_on="o_custkey")
-        .join(line_item_ds, left_on="o_orderkey", right_on="l_orderkey")
-        .filter(pl.col("o_orderdate") < var_2)
-        .filter(pl.col("l_shipdate") > var_1)
-        .with_columns(
-            (pl.col("l_extendedprice") * (1 - pl.col("l_discount"))).alias("revenue")
-        )
-        .group_by(["o_orderkey", "o_orderdate", "o_shippriority"])
-        .agg([pl.sum("revenue")])
-        .select(
-            [
-                pl.col("o_orderkey").alias("l_orderkey"),
-                "revenue",
-                "o_orderdate",
-                "o_shippriority",
-            ]
-        )
-        .sort(by=["revenue", "o_orderdate"], descending=[True, False])
-        .limit(10)
-    )
