--- tests/benchmarks/tpch/polars_queries/q5.py
@@ -1,38 +1,0 @@
-from datetime import datetime
-
-import polars as pl
-
-from .utils import read_data
-
-
-def query():
-    var_1 = "ASIA"
-    var_2 = datetime(1994, 1, 1)
-    var_3 = datetime(1995, 1, 1)
-
-    region_ds = read_data("region")
-    nation_ds = read_data("nation")
-    customer_ds = read_data("customer")
-    line_item_ds = read_data("lineitem")
-    orders_ds = read_data("orders")
-    supplier_ds = read_data("supplier")
-
-    return (
-        region_ds.join(nation_ds, left_on="r_regionkey", right_on="n_regionkey")
-        .join(customer_ds, left_on="n_nationkey", right_on="c_nationkey")
-        .join(orders_ds, left_on="c_custkey", right_on="o_custkey")
-        .join(line_item_ds, left_on="o_orderkey", right_on="l_orderkey")
-        .join(
-            supplier_ds,
-            left_on=["l_suppkey", "n_nationkey"],
-            right_on=["s_suppkey", "s_nationkey"],
-        )
-        .filter(pl.col("r_name") == var_1)
-        .filter(pl.col("o_orderdate").is_between(var_2, var_3, closed="left"))
-        .with_columns(
-            (pl.col("l_extendedprice") * (1 - pl.col("l_discount"))).alias("revenue")
-        )
-        .group_by("n_name")
-        .agg([pl.sum("revenue")])
-        .sort(by="revenue", descending=True)
-    )
