--- tests/benchmarks/tpch/polars_queries/q6.py
@@ -1,25 +1,0 @@
-from datetime import datetime
-
-import polars as pl
-
-from .utils import read_data
-
-
-def query():
-    var_1 = datetime(1994, 1, 1)
-    var_2 = datetime(1995, 1, 1)
-    var_3 = 24
-
-    line_item_ds = read_data("lineitem")
-
-    return (
-        line_item_ds.filter(
-            pl.col("l_shipdate").is_between(var_1, var_2, closed="left")
-        )
-        .filter(pl.col("l_discount").is_between(0.05, 0.07))
-        .filter(pl.col("l_quantity") < var_3)
-        .with_columns(
-            (pl.col("l_extendedprice") * pl.col("l_discount")).alias("revenue")
-        )
-        .select(pl.sum("revenue").alias("revenue"))
-    )
